requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
backcall=0.2.0=pypi_0
branca=0.4.1=pypi_0
ca-certificates=2020.11.8=ha878542_0
certifi=2020.11.8=py36h5fab9bb_0
click=7.1.2=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.17=pypi_0
datetime=4.3=pypi_0
decorator=4.4.2=pypi_0
fiona=1.8.17=pypi_0
folium=0.11.0=pypi_0
geopandas=0.8.1=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.35.1=hed1e6ac_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgomp=9.3.0=h5dbcf3e_17
libstdcxx-ng=9.3.0=h2ae2ef3_17
markupsafe=1.1.1=pypi_0
matplotlib=3.3.3=pypi_0
munch=2.5.0=pypi_0
ncurses=6.2=h58526e2_3
numpy=1.19.4=pypi_0
openssl=1.1.1h=h516909a_0
parso=0.7.1=pypi_0
patsy=0.5.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=pypi_0
pip=20.2.4=py_0
plotly=4.12.0=pypi_0
pmdarima=1.7.1=pypi_0
prompt-toolkit=3.0.8=pypi_0
ptyprocess=0.6.0=pypi_0
pygments=2.7.2=pypi_0
pyparsing=2.4.7=pypi_0
pyproj=3.0.0.post1=pypi_0
python=3.6.9=h9d8adfe_0_cpython
python_abi=3.6=1_cp36m
readline=8.0=he28a2e2_2
retrying=1.3.3=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.0=pypi_0
setuptools=49.6.0=py36h9880bd3_2
shapely=1.7.1=pypi_0
six=1.15.0=pypi_0
sqlite=3.33.0=h4cf870e_1
statsmodels=0.11.1=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hed695b0_1
traitlets=4.3.3=pypi_0
vadersentiment=3.3.2=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
zope-interface=5.2.0=pypi_0
